fix: Identifier class has circular reference and from method does not work
