+++ .changeset/warm-steaks-marry.md
@@ -1,0 +1,5 @@
+---
+'@masknet/base': patch
+---
+
+fix: Identifier class has circular reference and from method does not work
