+++ packages/base/src/Identifier/index.ts
@@ -1,6 +1,2 @@
-export * from './base.js'
-export * from './profile.js'
-export * from './post.js'
-export * from './post-iv.js'
-export * from './ec-key.js'
+export * from './identifier.js'
 export * from './map.js'
