+++ packages/base/src/Identifier/map.ts
@@ -1,8 +1,10 @@
-import { Identifier } from './base.js'
-import { ProfileIdentifier as A } from './profile.js'
-import { ECKeyIdentifier as B } from './ec-key.js'
-import { PostIdentifier as C } from './post.js'
-import { PostIVIdentifier as D } from './post-iv.js'
+import {
+    Identifier,
+    ProfileIdentifier as A,
+    ECKeyIdentifier as B,
+    PostIdentifier as C,
+    PostIVIdentifier as D,
+} from './identifier.js'
 
 type I<T> = Iterable<readonly [string, T]>
